Get PCA Data

This function computes K functions from PP data then converts it into PCs. Note, if there are repeated measures, i.e. multiple images per unit, the K functions will be (weighted) averaged together based on the number of agent1s.

Usage

getPCAData(
  data,
  outcome = colnames(data)[1],
  unit = colnames(data)[5],
  repeatedUniqueId = NULL,
  rCheckVals = NULL,
  nPCs = 3,
  agents_df = as.data.frame(expand.grid(unique(data[, 4]), unique(data[, 4]))),
  xRange = NULL,
  yRange = NULL,
  edgeCorrection = "isotropic",
  nbasis = 21,
  silent = FALSE,
  displayTVE = FALSE
)

Arguments

data

Dataframe with column titles for at least outcome, x, y, agents, and unit. For consistency (and avoiding errors), use that order. Additionally, repeatedUniqueId can be added.

outcome

(Optional) String of the column name in data indicating the outcome or response. Default is the 1st column.

unit

(Optional) String of the column name in data indicating a unit or base thing. Note this unit may have repeated measures. Default is the 4th column.

repeatedUniqueId

(optional) String of the column name in data indicating the unique ID when using repeated measures.

rCheckVals

(optional) numeric vector indicating the radius to check. Note, if note specified, this could take a lot of memory, particularly with many units and repeated measures.

nPCs

(optional) Numeric indicating the number of principal components

agents_df

(optional) Two-column data.frame. The first for agents 1 and the second for agent 2. Both should be in data agents column. This determines which K functions to compute. Default is to compute all, but may be misspecified if data in different order.

xRange

(optional) two value numeic vector indicating the min and max x values. Note this is re-used for all images. The default just takes the min and max x from each image. This allows different sized images, but note that the edges are defined by some cell.

yRange

(optional) two value numeic vector indicating the min and max y values. Note this is re-used for all images. The default just takes the min and max y from each image. This allows different sized images, but note that the edges are defined by some cell.

edgeCorrection

(optional) String indicating type of edgeCorrection to use in spatStat for computation of K functions.

nbasis

(optional) Numeric indicating number of basis functions to fit K functions in order to compute PCA.

silent

(optional) Boolean indicating if progress should be printed

displayTVE

(Optional) Boolean to indicate if total variance explained (TvE) should be displayed. Default is FALSE.

Value

Data.frame with the outcome, unit and principle components of computed K functions.

Examples

if (FALSE) {
data <- simulatePP(
  cellVarData =
    data.frame(
      "stage" = c(0, 1),
      "A" = c(0, 0),
      "B" = c(1 / 50, 1 / 50)
    ),
  cellKappaData = data.frame(
    "cell" = c("A", "B"),
    "clusterCell" = c(NA, "A"),
    "kappa" = c(20, 5)
  ),
  peoplePerStage = 100,
  imagesPerPerson = 1,
  silent = FALSE
)
agents_df_tmp <- as.data.frame(expand.grid(
  unique(data$cellType),
  unique(data$cellType),
  stringsAsFactors = FALSE
))
dat_pca <- getPCAData(
  data = data, outcome = "Stage", unit = "Person",
  repeatedUniqueId = "Image",
  rCheckVals = seq(0, 0.25, 0.01), nPCs = 3,
  agents_df = agents_df_tmp,
  xRange = c(0, 1), yRange = c(0, 1)
)
}

dataPCA_pheno <- getPCAData(
  data = TNBC_pheno, unit = "Person",
  agents_df = data.frame(rep("B",2), c("Tumour","FAKE")),
  nPCs = 3,
  rCheckVals = seq(0, 50, 1),
  displayTVE = TRUE
)
#> PCA Pairs (2): 1 (TVE: 0.997), 2